Bug fixes + shift click functionality
all known bugs have been squashed!
  - items stack properly
  - items don't disappear

shift + left click on an item slot to move the items in that slot and combine them with a different slot that also contains that item. if no other slots are found containing the same item, it will move to the first empty slot in the inventory.
this functionality (combining or moving to first empty slot) will also be integrated with mob drops post-merge.
to do:
- add final icons for all objects
- mob drops going to inventory upon death